[Reorg] Move settings to ui/settingspane
