+++ src/gl/controllers.cpp
@@ -31,7 +31,6 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ***** END LICENCE BLOCK *****/
 
 #include "controllers.h"
-#include "settings.h"
 
 #include "glmesh.h"
 #include "glnode.h"
