+++ src/gl/gltex.cpp
@@ -31,7 +31,6 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ***** END LICENCE BLOCK *****/
 
 #include "gltex.h"
-#include "settings.h"
 
 #include "glscene.h"
 #include "gltexloaders.h"
