+++ src/model/basemodel.h
@@ -45,6 +45,8 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <QVariant>
 #include <QVector>
 
+#define NifSkopeDisplayRole (Qt::UserRole + 42)
+
 
 //! @file basemodel.h BaseModel, BaseModelEval
