+++ src/model/nifdelegate.cpp
@@ -30,8 +30,6 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 ***** END LICENCE BLOCK *****/
 
-#include "settings.h"
-
 #include "glview.h"
 #include "kfmmodel.h"
 #include "nifmodel.h"
