+++ src/nifskope_ui.cpp
@@ -32,7 +32,6 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "nifskope.h"
 #include "version.h"
-#include "settings.h"
 
 #include "ui_nifskope.h"
 #include "ui/about_dialog.h"
